Rename NewContainer as BoxedContainer and the original BozedContainer as OldBoxedContainer.
