+++ .idea/workspace.xml
@@ -4,12 +4,15 @@
     <option name="autoReloadType" value="NONE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="398c2f1d-d59b-4632-b499-7ce03f824c85" name="Changes" comment="Implement widget.child null check to avoid zero or negative box constraints.">
+    <list default="true" id="398c2f1d-d59b-4632-b499-7ce03f824c85" name="Changes" comment="Movement of tiles on settings page fixed with NewContainer.&#10;&#10;NewContainer is a temporary new name for BBoxedContainer to differentiate it from (old) BoxedContainer in order to find and replace all instances of the old type.">
+      <change afterPath="$PROJECT_DIR$/lib/utils/old_boxed_container.dart" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/app_data/app_data.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/app_data/app_data.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/base_page/base_page.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/base_page/base_page.dart" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/lib/base_page/base_page_button_array.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/base_page/base_page_button_array.dart" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/lib/settings/settings_page_contents.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/settings/settings_page_contents.dart" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/lib/settings/settings_page_list_tile.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/settings/settings_page_list_tile.dart" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/lib/utils/bboxed_container.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/utils/bboxed_container.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/utils/bboxed_container.dart" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/utils/boxed_container.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/utils/boxed_container.dart" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -230,7 +233,14 @@
       <option name="project" value="LOCAL" />
       <updated>1690097553846</updated>
     </task>
-    <option name="localTasksCounter" value="24" />
+    <task id="LOCAL-00024" summary="Movement of tiles on settings page fixed with NewContainer.&#10;&#10;NewContainer is a temporary new name for BBoxedContainer to differentiate it from (old) BoxedContainer in order to find and replace all instances of the old type.">
+      <created>1690099884207</created>
+      <option name="number" value="00024" />
+      <option name="presentableId" value="LOCAL-00024" />
+      <option name="project" value="LOCAL" />
+      <updated>1690099884207</updated>
+    </task>
+    <option name="localTasksCounter" value="25" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -268,7 +278,8 @@
     <MESSAGE value="Address offset caused by adjusting childRect to account for borderWidth in _BBoxedContainerState." />
     <MESSAGE value="Update BBoxedContainer to account for a very small child.&#10;&#10;Check nullable variables." />
     <MESSAGE value="Implement widget.child null check to avoid zero or negative box constraints." />
-    <option name="LAST_COMMIT_MESSAGE" value="Implement widget.child null check to avoid zero or negative box constraints." />
+    <MESSAGE value="Movement of tiles on settings page fixed with NewContainer.&#10;&#10;NewContainer is a temporary new name for BBoxedContainer to differentiate it from (old) BoxedContainer in order to find and replace all instances of the old type." />
+    <option name="LAST_COMMIT_MESSAGE" value="Movement of tiles on settings page fixed with NewContainer.&#10;&#10;NewContainer is a temporary new name for BBoxedContainer to differentiate it from (old) BoxedContainer in order to find and replace all instances of the old type." />
   </component>
   <component name="XDebuggerManager">
     <breakpoint-manager>
@@ -309,22 +320,22 @@
           <option name="timeStamp" value="17" />
         </line-breakpoint>
         <line-breakpoint type="Dart">
-          <url>file://$PROJECT_DIR$/lib/utils/boxed_container.dart</url>
+          <url>file://$PROJECT_DIR$/lib/utils/old_boxed_container.dart</url>
           <line>260</line>
           <option name="timeStamp" value="21" />
         </line-breakpoint>
         <line-breakpoint type="Dart">
-          <url>file://$PROJECT_DIR$/lib/utils/boxed_container.dart</url>
+          <url>file://$PROJECT_DIR$/lib/utils/old_boxed_container.dart</url>
           <line>212</line>
           <option name="timeStamp" value="25" />
         </line-breakpoint>
         <line-breakpoint enabled="true" type="Dart">
-          <url>file://$PROJECT_DIR$/lib/utils/boxed_container.dart</url>
+          <url>file://$PROJECT_DIR$/lib/utils/old_boxed_container.dart</url>
           <line>221</line>
           <option name="timeStamp" value="27" />
         </line-breakpoint>
         <line-breakpoint enabled="true" type="Dart">
-          <url>file://$PROJECT_DIR$/lib/utils/boxed_container.dart</url>
+          <url>file://$PROJECT_DIR$/lib/utils/old_boxed_container.dart</url>
           <line>234</line>
           <option name="timeStamp" value="29" />
         </line-breakpoint>
