+++ lib/utils/boxed_container.dart
@@ -1,32 +1,20 @@
 // Import flutter packages.
-import 'dart:developer';
 import 'package:flutter/material.dart';
 import 'package:get_it_mixin/get_it_mixin.dart';
 
 // Import project-specific files.
 import 'package:kar_kam/app_data/app_data.dart';
-import 'package:kar_kam/utils/get_it_service.dart';
 import 'package:kar_kam/utils/global_key_extension.dart';
 
-
-
-
-
-
 /// Implements a [Container] and draws its bounding box.
-///
-/// [BoxedContainer] essentially calls an instance of [Container] with
-/// a default decoration if one is not specifically given. [BoxedContainer]
-/// defaults to [Container] if [AppData.drawLayoutBounds] is false.
 class BoxedContainer extends StatelessWidget with GetItMixin {
   BoxedContainer({
-    Key? key,
+    super.key,
     this.alignment,
-    this.borderColor,
-    this.borderRadius,
-    this.borderWidth,
+    this.borderColor = Colors.black,
+    this.borderWidth = 1.0,
     this.child,
-    this.clipBehavior = Clip.none,
+    this.clipBehavior,
     this.color,
     this.constraints,
     this.decoration,
@@ -38,12 +26,12 @@ class BoxedContainer extends StatelessWidget with GetItMixin {
     this.transform,
     this.transformAlignment,
     this.width,
-  }) : super(key: key);
+  });
 
   // [Container]-specific variables.
   final AlignmentGeometry? alignment;
   final Widget? child;
-  final Clip clipBehavior;
+  final Clip? clipBehavior;
   final Color? color;
   final BoxConstraints? constraints;
   final Decoration? decoration;
@@ -55,41 +43,32 @@ class BoxedContainer extends StatelessWidget with GetItMixin {
   final AlignmentGeometry? transformAlignment;
   final double? width;
 
-  // [BoxedContainer]-specific variables.
-  final Color? borderColor;
-  final double? borderRadius;
-  final double? borderWidth;
+  // [NewContainer]-specific variables.
+  final Color borderColor;
+  final double borderWidth;
   final bool drawLayoutBoundsOverride;
 
   @override
   Widget build(BuildContext context) {
     // Watch for changes to [AppData.drawLayoutBounds] registered with GetIt.
-    bool? drawLayoutBounds = watchOnly((AppData a) => a.drawLayoutBounds);
+    bool? drawLayoutBounds =
+        watchOnly((AppData a) => a.drawLayoutBounds) ?? false;
 
-    // Switch control of layout bounds from drawLayoutBounds, which is intended
-    // to be a global app setting, to borderWidth which is localised at the
-    // point of instantiation.
-    if (drawLayoutBoundsOverride && borderWidth != null) {
-      drawLayoutBounds = (borderWidth! > 0.0 ? true : false);
-    }
+    // Switch control of layout bounds from [drawLayoutBounds], which is
+    // intended to be a global app setting, to [drawLayoutBoundsOverride]
+    // which is localised at the point of instantiation.
+    drawLayoutBounds = drawLayoutBounds || drawLayoutBoundsOverride;
 
-    return Container(
+    return _BoxedContainer(
+      key: UniqueKey(),
       alignment: alignment,
+      borderColor: borderColor,
+      borderWidth: borderWidth,
       clipBehavior: clipBehavior,
+      color: color,
       constraints: constraints,
-      decoration: decoration ??
-          BoxDecoration(
-            border: drawLayoutBounds!
-                ? Border.all(
-              width: borderWidth ?? 0.1,
-              color: borderColor ?? Colors.black,
-            )
-                : null,
-            borderRadius: BorderRadius.all(
-              Radius.circular(borderRadius ?? 0.0),
-            ),
-            color: color,
-          ),
+      decoration: decoration,
+      drawLayoutBounds: drawLayoutBounds,
       foregroundDecoration: foregroundDecoration,
       height: height,
       margin: margin,
@@ -102,46 +81,18 @@ class BoxedContainer extends StatelessWidget with GetItMixin {
   }
 }
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-/// Implements a [Container] and draws its bounding box.
-///
-/// [BoxedContainer2] essentially calls an instance of [Container] with
-/// a default decoration if one is not specifically given. [BoxedContainer2]
-/// defaults to [Container] if [AppData.drawLayoutBounds] is false.
-class BoxedContainer2 extends StatefulWidget with GetItStatefulWidgetMixin {
-  BoxedContainer2({
+class _BoxedContainer extends StatefulWidget {
+  const _BoxedContainer({
     Key? key,
     this.alignment,
-    this.borderColor = Colors.black,
-    // this.borderRadius,
-    this.borderWidth = 0.1,
+    required this.borderColor,
+    required this.borderWidth,
     this.child,
-    this.clipBehavior = Clip.none,
+    this.clipBehavior,
     this.color,
     this.constraints,
     this.decoration,
+    required this.drawLayoutBounds,
     this.foregroundDecoration,
     this.height,
     this.margin,
@@ -154,7 +105,7 @@ class BoxedContainer2 extends StatefulWidget with GetItStatefulWidgetMixin {
   // [Container]-specific variables.
   final AlignmentGeometry? alignment;
   final Widget? child;
-  final Clip clipBehavior;
+  final Clip? clipBehavior;
   final Color? color;
   final BoxConstraints? constraints;
   final Decoration? decoration;
@@ -166,127 +117,104 @@ class BoxedContainer2 extends StatefulWidget with GetItStatefulWidgetMixin {
   final AlignmentGeometry? transformAlignment;
   final double? width;
 
-  // [BoxedContainer2]-specific variables.
+  // [NewContainer]-specific variables.
   final Color borderColor;
-  // final double? borderRadius;
   final double borderWidth;
+  final bool drawLayoutBounds;
 
   @override
-  State<BoxedContainer2> createState() => _BoxedContainer2State();
+  State<_BoxedContainer> createState() => _BoxedContainerState();
 }
 
-class _BoxedContainer2State extends State<BoxedContainer2> with GetItStateMixin {
-  // Required for establishing the bounding box for [Container] in [build].
-  final GlobalKey containerKey = GlobalKey();
+class _BoxedContainerState extends State<_BoxedContainer> {
+  // An [OverlayEntry] that will ultimately contain just a border to
+  // represent the layout bounds for [widget.child].
+  OverlayEntry? border;
 
-  // When evaluated, [containerRect] represents the layout bounds of the
-  // [Container] in [build]. [containerRect] is required for positioning
-  // the overlay.
-  Rect? containerRect;
+  // Used by [addBorder] for determining the bounding box for [widget.child].
+  final GlobalKey childKey = GlobalKey();
 
-  // Determines whether to draw layout bounds or not; set by initState.
-  bool? drawLayoutBounds = GetItService.instance<AppData>().drawLayoutBounds;
+  // // The link which connects the layers associated with
+  // LayerLink layerLink = LayerLink();
 
-  bool? drawLayoutBoundsOld;
+  // An object that displays [border] if [drawLayoutBounds] is true.
+  OverlayState? overlayState;
 
-  // The link which connects the layers associated with
-  // [CompositedTransformTarget] and [CompositedTransformFollower].
-  LayerLink layerLink = LayerLink();
+  // Generates the layout bounds for [widget.child].
+  void addBorder(Rect? rect) {
+    // Start with [border] as null.
+    removeBorder();
 
-  // Controlled by [drawLayoutBounds], this variable controls [OverlayPortal]
-  // in [build], which either shows or hides layout bounds.
-  OverlayPortalController overlayPortalController = OverlayPortalController();
+    // Create [border].
+    if (rect is Rect) {
+      // Adjust [rect] to account for [widget.borderWidth].
+      Rect borderRect = rect;//.inflate(0* -2 * widget.borderWidth);
+
+      // Avoid creating [border] if [borderRect] dimensions are too small.
+      if (borderRect.shortestSide < 2) return;
+
+      // Create [border].
+      border = OverlayEntry(
+        builder: (BuildContext context) {
+          return Stack(
+            children: <Widget>[
+              Positioned.fromRect(
+                rect: rect,
+                child: IgnorePointer(
+                  ignoring: true,
+                  child: Container(
+                    decoration: BoxDecoration(
+                      border: Border.all(
+                        width: widget.borderWidth,
+                        color: widget.borderColor,
+                      ),
+                      color: widget.color,
+                    ),
+                    height: borderRect.height,
+                    width: borderRect.width,
+                  ),
+                ),
+              ),
+            ],
+          );
+        },
+      );
+
+      // Insert the [border] OverlayEntry
+      overlayState = Overlay.of(context);
+      overlayState?.insert(border!);
+    }
+  }
+
+  // Make sure [border] is removed when the widget is disposed.
+  @override
+  void dispose() {
+    removeBorder();
+    super.dispose();
+  }
+
+  // Remove [border] and set to null.
+  void removeBorder() {
+    border?.remove();
+    border = null;
+  }
 
   @override
   void initState() {
-    // Set initial state for [overlayPortalController] whilst
-    // [Container] is built for the first time.
-    // overlayPortalController.hide();
-
-    // [_BoxedContainer2State] is built in two phases:
-    //    (i) with [Container], using [containerKey], and then
-    //    (ii) with [Container] and [OverlayPortal] showing the bounding box
-    //    if requested.
     WidgetsBinding.instance.addPostFrameCallback((_) {
-      setState(() {
-        // log('_BoxedContainer2State, initState...containerRect = $containerRect');
-        containerRect ??= containerKey.globalPaintBounds;
-        // log('_BoxedContainer2State, initState...containerRect = $containerRect');
-      });
+      // Get [child] bounding box characteristics if requested.
+      if (widget.drawLayoutBounds && widget.child != null) {
+        addBorder(childKey.globalPaintBounds);
+      }
     });
     super.initState();
   }
 
-  /// Set [overlayPortalController].
-  void setOverlayPortalController(bool bool) {
-    log('_BoxedContainer2State, setOverlayPortalController...${overlayPortalController.isShowing}');
-    bool ? overlayPortalController?.show() : overlayPortalController?.hide();
-    log('_BoxedContainer2State, setOverlayPortalController...${overlayPortalController.isShowing}');
-  }
-
   @override
   Widget build(BuildContext context) {
-    // Watch for changes to [AppData.drawLayoutBounds] registered with GetIt.
-    drawLayoutBounds = watchOnly((AppData a) => a.drawLayoutBounds);
-    // log('_BoxedContainer2State, build...drawLayoutBounds = $drawLayoutBounds');
-    // log('_BoxedContainer2State, build...containerRect = $containerRect');
-
-    if (containerRect != null) {
-      assert(drawLayoutBounds is bool,
-          '_BoxedContainer2State, build...drawLayoutBounds is null');
-      // assert(drawLayoutBoundsOld is bool,
-      //     '_BoxedContainer2State, build...drawLayoutBoundsOld is null');
-      if (drawLayoutBounds != drawLayoutBoundsOld) {
-        drawLayoutBoundsOld = drawLayoutBounds;
-        WidgetsBinding.instance.addPostFrameCallback((_) {
-          setOverlayPortalController(drawLayoutBounds!);
-          setState(() {
-            // log('_BoxedContainer2State, build...containerRect = $containerRect');
-            // containerRect ??= containerKey.globalPaintBounds;
-            // log('_BoxedContainer2State, build...containerRect = $containerRect');
-          });
-        });
-      }
-      // setOverlayPortalController(drawLayoutBounds!);
-      // SchedulerBinding.instance.addPostFrameCallback((_) {
-      //   //yourcode
-      // });
-    }
-    return OverlayPortal(
-      controller: overlayPortalController,
-      overlayChildBuilder: (BuildContext context) {
-        assert(containerRect != null,
-            '_BoxedContainer2State, build...containerRect is null');
-        containerRect?.inflate(-widget.borderWidth);
-        return Container(
-          decoration: BoxDecoration(
-            border: Border.all(
-              width: widget.borderWidth,
-              color: widget.borderColor,
-            ),
-            // color: Colors.pink,
-          ),
-          height: containerRect?.height,
-          width: containerRect?.width,
-          child: SizedBox.square(dimension: 5),
-        );
-      },
-      child: Container(
-        key: containerKey,
-        alignment: widget.alignment,
-        clipBehavior: widget.clipBehavior,
-        color: widget.color,
-        constraints: widget.constraints,
-        decoration: widget.decoration,
-        foregroundDecoration: widget.foregroundDecoration,
-        height: widget.height,
-        margin: widget.margin,
-        padding: widget.padding,
-        transform: widget.transform,
-        transformAlignment: widget.transformAlignment,
-        width: widget.width,
-        child: widget.child,
-      ),
+    return Container(
+      key: childKey,
+      child: widget.child,
     );
   }
 }
